Add .emacs.moss

+++ setup_emacs.sh
@@ -1,11 +1,19 @@
 #!/bin/zsh
+
 CONFIG=~/gitws/config
 if ! [ -e $CONFIG -a -e `which git` ]; then
   echo "git clone first!"
   exit 1
 fi
+HOSTNAME=`hostname -s`
+EMACSNAME=$CONFIG/.emacs.customized/.emacs.$HOSTNAME
 
-! [ -e ~/.emacs ] && ln -s $CONFIG/.emacs ~/.emacs
+if [ -e $EMACSNAME ]; then
+    ln -s $EMACSNAME ~/.emacs
+else
+    ! [ -e ~/.emacs ] && ln -s $CONFIG/.emacs ~/.emacs
+fi
+
 ! [ -e ~/.emacs.d ] && ln -s $CONFIG/.emacs.d ~/.emacs.d
 ! [ -e ~/.emacs.d/elisp ] && mkdir ~/.emacs.d/elisp
